requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.13
distro==1.9.0
edge-tts==7.2.3
flatbuffers==25.9.23
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
idna==3.11
jiter==0.11.1
joblib==1.5.2
keyboard==0.13.5
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
numpy==2.3.4
onnxruntime==1.23.2
openai==2.7.1
openwakeword==0.6.0
packaging==25.0
pillow==12.0.0
playsound==1.2.2
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pvporcupine==3.0.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20251023
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyGetWindow==0.0.9
pyjokes==0.8.3
PyMsgBox==2.0.1
pynput==1.7.7
pyparsing==3.2.5
pyperclip==1.11.0
pypiwin32==223
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.8
SpeechRecognition==3.14.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wikipedia==1.4.0
yarl==1.22.0
